+++ extensions/omni/init.c
@@ -28,16 +28,22 @@ static shmem_startup_hook_type saved_shmem_startup_hook;
 static void shmem_request();
 static void shmem_hook();
 
-void procoid_syscache_callback(Datum arg, int cacheid, uint32 hashvalue) {
-  backend_force_reload = true;
-}
-
 extern void deinitialize_backend(int code, Datum arg);
+
+MODULE_VARIABLE(int ServerVersionNum);
 
 /**
  * Shared preload initialization.
  */
 void _PG_init() {
+  {
+    // Establish the presence marker
+    void **rendezvous = find_rendezvous_variable("omni(loaded)");
+    static struct _omni_rendezvous_var_t rendezvous_var = {.magic = "0MNI", .version = EXT_VERSION};
+    rendezvous_var.library_path = get_omni_library_name();
+    *rendezvous = &rendezvous_var;
+  }
+
   memset(saved_hooks, 0, sizeof(saved_hooks));
   // This signifies if this library has indeed been preloaded
   static bool preloaded = false;
@@ -144,10 +150,6 @@ void _PG_init() {
     RegisterBackgroundWorker(&master_worker);
   }
 
-  // This function may result in a FATAL error if we hit the limit of
-  // syscache callbacks (`MAX_SYSCACHE_CALLBACKS`, 64). However, since we're in
-  // postmaster: 1) it is okay to crash here 2) it is much less likely to happen at this time.
-  CacheRegisterSyscacheCallback(PROCOID, procoid_syscache_callback, Int8GetDatum(0));
   backend_force_reload = true;
 
   OmniGUCContext = AllocSetContextCreate(TopMemoryContext, "omni:guc", ALLOCSET_DEFAULT_SIZES);
@@ -156,6 +158,14 @@ void _PG_init() {
 
   xact_oneshot_callbacks = NIL;
   after_xact_oneshot_callbacks = NIL;
+
+  ServerVersionNum = pg_strtoint32(GetConfigOption("server_version_num", false, false));
+
+  if (ServerVersionNum != PG_VERSION_NUM) {
+    ereport(WARNING, errmsg("omni has been compiled against %d.%d, but running on %d.%d",
+                            PG_VERSION_NUM / 10000, PG_VERSION_NUM % 100, ServerVersionNum / 10000,
+                            ServerVersionNum % 100));
+  }
 }
 
 /**